requirements.txt
# Dépendances pour le système RAG avec scraping

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# RAG et LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0

# Base de données vectorielle
chromadb>=0.4.0

# Embeddings
openai>=1.0.0
sentence-transformers>=2.2.0
huggingface-hub>=0.20.0

# Gestion des variables d'environnement
python-dotenv>=1.0.0

# Utilitaires
tiktoken>=0.5.0
